refactor(telemetry): move CLI reporting helpers out of main into telemetry
main.rs now calls a single telemetry::record_invocation(matches, error,
duration) at exit. The clap-matches walk (command path, flags,
request_shape), failure classification, spool, and bounded/blocking
flush all live in the telemetry module. ReportContext, enqueue, and the
flush fns are now module-private; main no longer pulls in clap
ArgMatches/ValueSource or the engine error types for this.
